Earthquake visualisation exercise
Implement the following in Python.  Follow the same practices you would apply to production-standard code.

 

1)      Take a look at this site for the US Geological Service Rest API for earthquakes.  These APIs are publically available and free to use.- https://earthquake.usgs.gov/data/data.php#eq 

 

2)      Create a request that returns all earthquake data for Jan 2016  in geojson format

 

4)   Utilise OOPs concepts to create a suitable class definition for Earthquake.

 

5)      Call this request from Python and store the results in a suitable Python data structure.

 

6)      Extend the code to pick up all earthquake data for 2016 – note this cannot be done in a single call, as there is a limit of 20,000 events per request. Look at parallel request implementations to speed up data gathering.

 

7)      Use a suitable Python visualisation library e.g. matplotlib, to visualise the data. Use your creativity and see whether there are any conclusions you can draw from the data

 

8) Look to create unit test/s and suitable integration test/s to aid your codebase.

 

9) We would appreciate if you could share with us a github repo link which has your implementation.

10) Logging

 

 

Preferred Python version is 3.4 or above.

 

Notes – it is very easy to write Python code quickly that doesn’t work in a production situation.  In particular I will be looking for any techniques used to make the code of production quality.  I will also be looking for creativity in visualisation. Finally the development team use Pandas. If you are able to adapt to demonstrate use of Pandas Dataframes, that is a bonus.